backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pyswisseph==2.10.3.2
python-dotenv==1.0.0
openai==1.40.0
anthropic==0.34.0
langchain>=0.3.13
langchain-anthropic>=0.3.4
langchain-openai>=1.1.7
langchain-core>=0.3.31
supabase>=2.27.3
structlog>=25.5.0
